[dataset] emit OT_CHANGED_* events when applying Dataset for first time
This commit adds new logic in `Dataset::ApplyConfiguration()` to
ensure when a Dataset configuration is applied, each of the
`OT_CHANGED_*` events (for channel, network name, PAN id, or extended
PAN id) is emitted for the first time it is applied. Afterwards only
if the value is changed the corresponding `OT_CHANGED_*` is  emitted.

This change addresses an issue where if user sets any of these
parameters (using OpenThread public APIs) before starting Thread
operation, the event indicating the change would not be emitted.