[cc2538] avoid possible deadlock in otPlatRadioTransmit (#2371)
* [cc2538] Add definitions for FSMSTAT0 and RFERRF bits.

This adds a register definition for `RFCORE_XREG_FSMSTAT0` and defines
the bits in `RFCORE_SFR_RFERRF`.

* [cc2538] Check for error conditions on transmit.

The transmit code first waits for an existing transmission to finish,
then sets things up for the next transmission and waits for the radio to
go into receive mode before proceeding.

If, the radio receive buffer overflows during this time, the radio will
enter a RX overflow state (state 0x11 in `FSMSTAT0`) and will block
there until the RX FIFO is flushed.

In such cases, the CC2538 hard-locks due to the while loop waiting for
the `RX_ACTIVE` bit to be set.  While waiting for `RX_ACTIVE`, we need
to check for these overflow conditions, handle them, then re-try our
transmitter set-up.

* [cc2538] Re-try transmit set-up on IDLE

If we enter the idle state, then rather than just setting the channel,
re-try our transmitter set-up since we should be in the correct state
to do this at this point.