[core] add Set class
This commit adds a new common class `Set` which represents a set of
integers with a given (template) size parameter. The `Set` class is
implemented using a bit-vector mask array. This commit also adds a
unit test for the new `Set` class. The `ChannelMask`, `RouterTable`,
and `Ncp::ChangedPropsSet` classes are modified to use the new
`Set` class.