[mle-router] stop mStateUpdateTimer on mode change
This commit ensures to not re-start the `mStateUpdateTimer` from its
callback, when router role is disabled. This addresses the issue where
a mode change (disabling router role) could cause the timer to
continue to fire indefinitely, further causing device to attempt to
reattach from each timer callback.

Thanks to Conor O'Neill (@conormoneill) for discovering and debugging
this issue.