[cc2538] add support for CC2538-CC2592-EM (#3609)
* [cc2538] Define registers for control of CC2592.

In order to set up the signals needed to drive the CC2592 front-end, we
need to expose a few registers in the radio module that route internal
module signals to pins on Port C.

TI also recommend hooking the HGM pin of the CC2592 to a GPIO pin so
that the high-gain mode can be software-controlled.  It can be connected
to any GPIO pin, but they call out PD2 specifically for this, so we need
to configure GPIO port C & D.  Since we're here, we'll add B as well.

Some macros of convenience are also added for accessing the Nth element
in an array of 32-bit registers, as well as computing the bit mask of
the Nth GPIO pin.

* [cc2538] Define configuration settings for CC2592.

By default, we disable support for the CC2592 as we do not wish to break
existing OpenThread instances.  The configuration parameters allow
customisation of the locations for the `PA_EN`, `LNA_EN` and `HGM`
signals (within the constraints of the CC2538) for cases where people
have moved the signals around on custom boards.

The defaults work with designs derived off the CC2538-CC2592 EM
reference design.

* [cc2538] Define control interface for HGM pin.

Since the `HGM` signal is under software control, we need to be able to
allow the software to turn it on or off as required.  This exposes two
functions, one of which queries the current state (by looking at the
GPIO register bit-mask), and the other provides a means to change the
state.

* [cc2538] Implement control and calibration of the CC2592.

This implements the control functions defined earlier and adds the
necessary logic for initialising the CC2592.

The transmit and receive power levels are also calibrated.

* [cc2538] Define CC2592 options in Makefile.

This allows building of a CC2538-CC2592 EM image by specifying
`CC2592=1` on the command line.