[joiner] allow EUI64 address to be specified from Joiner::Start()
This commit extends the `Joiner::Start()` to allow a EUI64 address
to be (optionally) provided as an input parameter which is used to
derive the Joiner Id during the Joiner commissioning process. If a
`NULL` pointer is passed as the EUI64, the factory-assigned address
(from radio platform API `otPlatRadioGetIeeeEui64()`) will be used
instead. This commit also updates the public API `otJoninerStart()`
accordingly.