[posix-sim] use multicast to simulate radio
This PR changes to use multicast to simulate radio transmission. This
avoids sending a frame multiple times, and also avoid duplicated packets
in captures which is annoying when analyzing traffic.

This implementation uses two separate sockets for tx and rx. This is
because the current implementation of thread-cert requires sniffer
reports the exact sender of each packet. This issue can be addressed by
stateful tracking state of each node, but I don't have the cycles for
now.