[platform] Add hook for flushing the UART buffer.
This is used to allow the CLI to flush its transmit buffer of pending
data so that it can resume writing to the buffer.

This allows us to reduce the size of the transmit buffer used for the
CLI whilst still retaining the ability to transmit long blocks of text
(which is required by the Thread certification tests) without locking
up the use of that memory on constrained devices like CC2538.