[radio] use a single variable for frame timestamp
This PR remove mSec and mUsec from otRadioFrame, and use a single
mTimestamp to record when the frame is received.

The timestamp MUST be the time when SFD was received when TIME_SYNC or
CSL(Thread 1.2) is enabled, otherwise, it MAY be the time when rx was
done.

This change reduces memory required to represent a radio frame.