[cc2538] Allow RX interrupt to be disabled.
This was actually a whoopsie on my part, the affected lines were
erroniously added as part of the CC2592 support enhancement (commit
fd2503e792da9aa3bfd08a02bdb2b320936e3d6d) which had nothing to do with
CC2592, but was in our code because we *did* have an interrupt-driven
radio driver.

Leaving this turned on without actually properly implementing reception
is begging to drop frames without warning.  The next commits will fix
this, implementing a receive handler, but it may be desired to turn this
off for debugging, so let's use the configuration option defined
earlier.