[cc2538] Avoid debug logs in receiveFrame
If we call this method from an interrupt handler, we want to avoid
calling `snprintf` and similar functions as these are generally not
interrupt-safe.

Instead, use an extra byte of RAM to store the number of bytes dropped
so we can log it from a safe context.  This is only needed when both
interrupts _and_ platform logging are enabled.